test: fix tests for all packages
