+++ packages/app-degree-pages/__mocks__/data/degree-search-detail.json
@@ -13,6 +13,7 @@
         "WEST:REG:2227": "2022 Fall:null:2022-08-23:null: ",
         "ONLNE:SESB:2217": "2021 Fall:null:2021-10-13:null: "
       },
+      "Degree": "BA",
       "CollegeUrl": "https://wpcarey.asu.edu/",
       "EmailAddr": "wpcareyug@asu.edu",
       "CurriculumUrl": "https://asuonline.asu.edu/online-degree-programs/undergraduate/bachelor-arts-business-global-logistics-management/",
