+++ packages/app-degree-pages/cypress.config.js
@@ -1,0 +1,11 @@
+const { defineConfig } = require('cypress')
+
+module.exports = defineConfig({
+  e2e: {
+    setupNodeEvents(on, config) {
+      return require('./cypress/plugins/index.js')(on, config)
+    },
+    baseUrl: 'http://localhost:9010',
+    excludeSpecPattern: '**/examples/*.spec.js',
+  },
+})
