--- packages/app-degree-pages/cypress.json
@@ -1,5 +1,0 @@
-{
-	"baseUrl": "http://localhost:9009",
-	"$schema": "https://on.cypress.io/cypress.schema.json",
-	"ignoreTestFiles": "**/examples/*.spec.js"
-}
