+++ packages/app-degree-pages/cypress/jsconfig.json
@@ -1,5 +1,7 @@
 {
-  "target": "es2015",
+  "compilerOptions": {
+    "target": "es6"
+  },
   "include": [
     "./node_modules/cypress",
     "cypress/**/*.js",
