+++ packages/app-degree-pages/jest.config.js
@@ -7,6 +7,9 @@ module.exports = {
     "^.+\\.css$": "jest-transform-css",
     "\\.(jpg|jpeg|png|gif|webp|svg)$": "jest-transform-file",
   },
+  moduleNameMapper: {
+    "^@asu/components-core$": "<rootDir>/../components-core/src/index.js",
+  },
   transformIgnorePatterns: ["/node_modules/"],
   testPathIgnorePatterns: [
     "@asu/components-core",
