+++ packages/app-degree-pages/package.json
@@ -26,7 +26,7 @@
   },
   "scripts": {
     "lint": "eslint --fix 'src/**/*.{js,jsx}' --ignore-path ../../.eslintignore",
-    "test": "jest --config=./jest.config.js --passWithNoTests --coverage",
+    "test": "jest --config=./jest.config.js --silent --coverage",
     "test-update-snapshot": "yarn test -- -u",
     "e2e-ci": "concurrently --kill-others \"yarn storybook\" \"yarn cy:run\"",
     "e2e": "yarn cy:open",
@@ -44,7 +44,7 @@
     "postdocs": "node ../../scripts/process-readme-props.js"
   },
   "dependencies": {
-    "@asu/components-core": "^2.1.0",
+    "@asu/components-core": "^3.1.0",
     "@popperjs/core": "^2.9.2",
     "classnames": "^2.3.1",
     "prop-types": "^15.7.2",
