+++ packages/app-rfi/package.json
@@ -26,7 +26,7 @@
   },
   "scripts": {
     "lint": "eslint --fix 'src/**/*.{js,jsx}' --ignore-path ../../.eslintignore",
-    "test": "jest --config=./jest.config.js",
+    "test": "jest --config=./jest.config.js --passWithNoTests",
     "start:dev": "webpack-dashboard -- webpack serve -c webpack/webpack.dev.js",
     "prebuild": "rm -rf ./dist",
     "build": "webpack -c webpack/webpack.prod.js",
