+++ packages/component-carousel/jest.config.js
@@ -6,10 +6,10 @@ module.exports = {
     "^.+\\.js$": "babel-jest",
     "^.+\\.(js|jsx)$": "babel-jest",
     "^.+\\.css$": "jest-transform-css",
-    "^.+\\.scss$": "jest-transform-scss",
     "\\.(jpg|jpeg|png|gif|webp|svg)$": "jest-transform-file",
   },
   moduleNameMapper: {
     "\\.(css|less|sass|scss)$": "identity-obj-proxy",
+    "^@asu/components-core$": "<rootDir>/../components-core/src/index.js",
   },
 };
