+++ packages/component-cookie-consent/package.json
@@ -26,7 +26,7 @@
   },
   "scripts": {
     "lint": "eslint --fix 'src/**/*.{js,jsx}' --ignore-path ../../.eslintignore",
-    "test": "jest --config=./jest.config.js --silent --coverage",
+    "test": "jest --config=./jest.config.js --silent --coverage --passWithNoTests",
     "test-update-snapshot": "yarn test -- -u",
     "prebuild": "rm -rf ./dist",
     "build": "webpack -c webpack/webpack.prod.js",
