--- packages/unity-bootstrap-theme/__tests__/unity-bootstrap-theme.test.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const unityBootstrapTheme = require('..');
-const assert = require('assert').strict;
-
-assert.strictEqual(unityBootstrapTheme(), 'Hello from unityBootstrapTheme');
-console.info("unityBootstrapTheme tests passed");
